+++ src/modules/companies/services/update-company-logo.service.ts
@@ -38,8 +38,10 @@ export class UpdateCompanyLogoService {
 			);
 		}
 
-		const filename = foundCompany.logo.split('uploads/')[1];
-		fs.unlinkSync(`uploads/${filename}`);
+		if (foundCompany.logo) {
+			const filename = foundCompany.logo.split('uploads/')[1];
+			fs.unlinkSync(`uploads/${filename}`);
+		}
 
 		foundCompany.logo = `${process.env.BASE_URL}:${
 			process.env.PORT || 3333
